backend/requirements.txt
Django==3.2.3
djangorestframework==3.12.4
drf-extra-fields==3.4.0
drf-yasg==1.21.7
django-filter==2.4.0
django-extensions==3.1.3
djoser==2.1.0
gunicorn==20.1.0
psycopg2-binary==2.9.3
Pillow==11.1.0
requests==2.26.0
python-dotenv==1.0.1
